Move CustomAdapter SearchResultFragment.kt to ui.searchReesult
