+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/searchResult/CustomAdapter.kt
@@ -1,0 +1,39 @@
+package jp.co.yumemi.android.code_check.ui.searchResult
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.model.Repository
+
+class CustomAdapter(
+    private val itemClickListener: OnItemClickListener,
+) : ListAdapter<Repository, CustomAdapter.ViewHolder>(diff_util){
+
+    class ViewHolder(view: View): RecyclerView.ViewHolder(view)
+
+    interface OnItemClickListener{
+        fun itemClick(item: Repository)
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder
+    {
+        val _view= LayoutInflater.from(parent.context)
+            .inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(_view)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int)
+    {
+        val _repository= getItem(position)
+        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text=
+            _repository.name
+
+        holder.itemView.setOnClickListener{
+            itemClickListener.itemClick(_repository)
+        }
+    }
+}
